requirements.txt
PyYAML
slugify
colorama
requests
beautifulsoup4
